+++ actionview/lib/action_view/testing/resolvers.rb
@@ -41,7 +41,7 @@ module ActionView #:nodoc:
           )
         end
 
-        templates.sort_by { |t| -t.identifier.match(/^#{query}$/).captures.reject(&:blank?).size }
+        templates.sort_by { |t| -t.identifier.match(/^#{query}$/).captures.compact_blank.size }
       end
   end