+++ activerecord/lib/active_record/connection_adapters/connection_specification.rb
@@ -50,7 +50,7 @@ module ActiveRecord
 
         # Converts the given URL to a full connection hash.
         def to_hash
-          config = raw_config.reject { |_, value| value.blank? }
+          config = raw_config.compact_blank
           config.map { |key, value| config[key] = uri_parser.unescape(value) if value.is_a? String }
           config
         end